+++ src/compas_viewer/scene/lineobject.py
@@ -14,7 +14,9 @@ class LineObject(ViewerSceneObject, GeometryObject):
     """
 
     def __init__(self, line: Line, **kwargs):
-        super(LineObject, self).__init__(geometry=line, **kwargs)
+        if kwargs["show_lines"] is None:
+            kwargs["show_lines"] = True
+        super().__init__(geometry=line, **kwargs)
 
     def _read_points_data(self) -> DataType:
         positions = [self.geometry.start, self.geometry.end]