+++ src/compas_viewer/scene/pointobject.py
@@ -8,13 +8,22 @@ from .sceneobject import ViewerSceneObject
 class PointObject(ViewerSceneObject, GeometryObject):
     """Viewer scene object for displaying COMPAS Point geometry.
 
+    Parameters
+    ----------
+    point : :class:`compas.geometry.Point`
+        The point geometry to display.
+    show_points : bool, optional
+        Whether to display the point in the viewer. Default is True.
+
     See Also
     --------
     :class:`compas.geometry.Point`
     """
 
     def __init__(self, point: Point, **kwargs):
-        super(PointObject, self).__init__(geometry=point, **kwargs)
+        if kwargs["show_points"] is None:
+            kwargs["show_points"] = True
+        super().__init__(geometry=point, **kwargs)
         self.geometry: Point
 
     def _read_points_data(self) -> DataType: