+++ src/compas_viewer/scene/polygonobject.py
@@ -13,4 +13,4 @@ class PolygonObject(MeshObject):
     """
 
     def __init__(self, polygon: Polygon, **kwargs):
-        super(PolygonObject, self).__init__(mesh=Mesh.from_shape(polygon), **kwargs)
+        super().__init__(mesh=Mesh.from_shape(polygon), **kwargs)